debug

+++ fabric/activity/getrange/activity.go
@@ -237,7 +237,7 @@ func retrieveRange(ctx activity.Context, ccshim shim.ChaincodeStubInterface, sta
 		log.Debugf("result metatadata: %+v\n", *resultMetadata)
 		if resultMetadata != nil {
 			log.Debugf("set count %v and bookmark %s\n", resultMetadata.FetchedRecordsCount, resultMetadata.Bookmark)
-			ctx.SetOutput(ovCount, resultMetadata.FetchedRecordsCount)
+			ctx.SetOutput(ovCount, float64(resultMetadata.FetchedRecordsCount))
 			ctx.SetOutput(ovBookmark, resultMetadata.Bookmark)
 			c := ctx.GetOutput(ovCount)
 			log.Debugf("verify count: %T, %v\n", c, c)
